docker-compose/requirements/app/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
django-cors-headers